server/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
Pillow 